Moved FileSelector to Util package
